requirements.txt
cap2geojson
cryptography==46.0.3
eccodes==1.7.1
elasticsearch==7.16.2
elasticsearch-dsl==7.4.1
filelock
flask>=2.2.0
flask-cors
jsonpatch
json-merge-patch
paho-mqtt
pygeometa
pymetdecoder-wmo==0.1.14
synop2bufr==0.7.0
csv2bufr==0.8.6
minio==7.2.7
bufr2geojson==0.7.1
geopy==2.4.1
shapely==2.1.1
